feat: move numeric types into Candoumbe.Types.Numerics project
