+++ src/Candoumbe.Types.Numerics/Candoumbe.Types.Numerics.csproj
@@ -1,10 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
   <Import Project="..\..\core.props"/>
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
+  <PropertyGroup>
+    <TargetFrameworks>netstandard2.1;net8.0;</TargetFrameworks>
+  </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Primitives" Version="8.0.0" Condition="'$(TargetFramework)' == 'net8.0'"/>
+    <PackageReference Update="Candoumbe.MiscUtilities" Version="0.14.0"/>
+  </ItemGroup>
 </Project>
