+++ src/Candoumbe.Types.Numerics/Number.cs
@@ -1,0 +1,64 @@
+﻿using System;
+#if NET7_0_OR_GREATER
+using System.Numerics;
+#endif
+
+namespace Candoumbe.Types.Numerics;
+
+/// <summary>
+/// Base class for creating numeric types that can only hold a certain range of values.
+/// </summary>
+/// <typeparam name="TNumber">The numeric type that the current type will restrict values for.</typeparam>
+public abstract record Number<TNumber> : IComparable<Number<TNumber>>
+    where TNumber : IComparable<TNumber>
+{
+    /// <summary>
+    /// Gets the underlying value
+    /// </summary>
+    /// <remarks>The value is guaranteed to be within the range of values defined by the current type.</remarks>
+    public TNumber Value
+    {
+        get;
+#if NET7_0_OR_GREATER
+        init;
+#else
+            protected set;
+#endif
+    } = default;
+
+    /// <summary>
+    /// Builds a new <see cref="Number{TNumber}"/> that contains the default value of <typeparamref name="TNumber"/>.
+    /// </summary>
+    protected Number() => Value = default;
+
+    /// <summary>
+    /// Builds a new <see cref="Number{TNumber}"/> initialized with the specified <paramref name="value"/>.
+    /// </summary>
+    /// <param name="value"></param>
+    protected Number(TNumber value) => Value = value;
+
+    ///<inheritdoc/>
+    public int CompareTo(Number<TNumber> other) => Value.CompareTo(other.Value);
+
+    ///<inheritdoc/>
+    public virtual bool Equals(Number<TNumber> other) => Equals(Value, other.Value);
+
+    ///<inheritdoc/>
+    public override int GetHashCode() => Value.GetHashCode();
+
+    ///<inheritdoc/>
+    public static bool operator <(Number<TNumber> left, TNumber right)
+        => left.Value.CompareTo(right) < 0;
+
+    ///<inheritdoc/>
+    public static bool operator >(Number<TNumber> left, TNumber right)
+        => left.Value.CompareTo(right) > 0;
+
+    ///<inheritdoc/>
+    public static bool operator >=(Number<TNumber> left, TNumber right)
+        => left.Value.CompareTo(right) >= 0;
+
+    ///<inheritdoc/>
+    public static bool operator <=(Number<TNumber> left, TNumber right)
+        => left.Value.CompareTo(right) <= 0;
+}
