+++ test/Candoumbe.Types.Numerics.UnitTests/Candoumbe.Types.Numerics.UnitTests.csproj
@@ -14,6 +14,4 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-
-
 </Project>
