+++ test/Candoumbe.Types.Numerics.UnitTests/Generators/ValueGenerators.cs
@@ -26,4 +26,10 @@ internal static class ValueGenerators
                                                                          .Generator
                                                                          .Select(value => NonNegativeInteger.From(value.Item))
         .ToArbitrary();
+    
+    public static Arbitrary<NonNegativeLong> NonNegativeLongs() => ArbMap.Default.ArbFor<long>()
+        .Filter(value => value >= 0)
+        .Generator
+        .Select(NonNegativeLong.From)
+        .ToArbitrary();
 }
