+++ test/Candoumbe.Types.UnitTests/Candoumbe.Types.UnitTests.csproj
@@ -14,8 +14,4 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Calendar\" />
-  </ItemGroup>
 </Project>
